~ Wellformedness_checks refactoring
