+++ server/src/symbol_table/create_symbol_table.ts
@@ -1,7 +1,7 @@
 import { TextDocument } from "vscode-languageserver-textdocument";
 import Parser = require("web-tree-sitter");
 import { Range, Position } from 'vscode-languageserver/node';
-import { check_reserved_facts, checks_with_table } from '../features/wellformedness_checks';
+import { check_reserved_facts} from '../features/checks/checkReservedFacts';
 import {getName} from '../features/syntax_errors'
 import { Diagnostic } from "vscode-languageserver-types";
 import path = require('path');
